add indentity membership system add login and registration functional update nuget packages

+++ MvcTomato/Controllers/HomeController.cs
@@ -5,16 +5,21 @@ using System.Linq;
 using System.Web.Mvc;
 using MvcTomato.DAL;
 using MvcTomato.Models;
-using MvcTomato.ViewModels;
 using HistoryDayViewModel = MvcTomato.ViewModels.HistoryDayViewModel;
+using Microsoft.AspNet.Identity;
+using MvcTomato.Utils;
+using MvcTomato.ViewModels;
 
 /* 
 * Implement ability saving partial data (e.g. only enter and dinner start dates)
 * And show them on the next page loading (user log in)
+* TODO: add logging NLog + interface
+* TODO: unit tests for controllers (moq, IoC)
 */
 
 namespace MvcTomato.Controllers
 {
+    [Authorize]
     public class HomeController : Controller
     {
         private TomatoContext db = new TomatoContext();
@@ -22,11 +27,20 @@ namespace MvcTomato.Controllers
         public ActionResult Index()
         {
             TimeSpan rate = TimeSpan.FromHours(7.5);
-            var month = db.WorkingDays.Where(d => d.Date.Month == DateTime.Today.Month && d.Finished).ToList();
+            string userId = User.Identity.GetUserId();
+            var month = db.WorkingDays
+                .Where(d => d.OwnerId == userId)
+                .Where(d => d.Date.Month == DateTime.Today.Month && d.Finished)
+                .ToList();
+
             var monthStat = month.OrderBy(d => d.Date).Select(d => rate - (d.Exit - d.Enter - (d.DinnerFinish - d.DinnerStart))).ToList();
             // Need to separate these two querys because substruction of dates is not supported in SQL
             // I convet it to List in order to force query execution (or maybe DbFunctions can be used)
-            var today = db.WorkingDays.Where(d => d.Date == DateTime.Today && d.Finished).ToList();
+            var today = db.WorkingDays
+                .Where(d => d.OwnerId == userId)
+                .Where(d => d.Date == DateTime.Today && d.Finished)
+                .ToList();
+
             var todayStat = new TimeSpan?();
             if (today.Any())
             {
@@ -37,35 +51,39 @@ namespace MvcTomato.Controllers
             ViewBag.MonthAllSum = month.Select(d => d.Exit - d.Enter - (d.DinnerFinish - d.DinnerStart)).Aggregate(TimeSpan.Zero, (TimeSpan? t1, TimeSpan? t2) => t1 + t2);
             ViewBag.DayStatistics = todayStat;
 
-            var uncompleteToday = db.WorkingDays.FirstOrDefault(d => d.Date == DateTime.Today && !d.Finished);
+            var uncompleteToday = db.WorkingDays
+                .Where(d => d.OwnerId == userId)
+                .FirstOrDefault(d => d.Date == DateTime.Today && !d.Finished);
             // TODO: Add other uncomplete days
             if (uncompleteToday != null)
             {
-                ViewBag.UncompleteToday = uncompleteToday;
+                ViewBag.UncompleteToday = ModelConverter.ToViewModel(uncompleteToday);
             }
             return View();
         }
 
         [HttpPost]
-        public ActionResult Index(WorkingDay day)
+        public ActionResult Index(WorkingDayViewModel day)
         {
-            System.Diagnostics.Debug.Print("Test");
+            string userId = User.Identity.GetUserId();
+            var dayModel = ModelConverter.ToModel(day);
+            dayModel.OwnerId = userId;
             if (ModelState.IsValid)
             {
                 // TODO: Implement it using AJAX
                 // Check there is no entrys in DB with that date
-                if (db.WorkingDays.Any(d => d.Date == day.Date))
+                if (db.WorkingDays.Where(d => d.OwnerId == userId).Any(d => d.Date == dayModel.Date))
                 {
                     return View("Error");
                 }
                 // TODO: for debug purpose
-                if (day.Enter != null && day.Exit != null && day.DinnerStart != null && day.DinnerFinish != null)
+                if (dayModel.Enter != null && dayModel.Exit != null && dayModel.DinnerStart != null && dayModel.DinnerFinish != null)
                 {
-                    day.Finished = true;
+                    dayModel.Finished = true;
                 }
-                db.WorkingDays.Add(day);
+                db.WorkingDays.Add(dayModel);
                 db.SaveChanges();
-                return View();
+                return RedirectToAction("Index", "Home");
             }
             return View(day);
         }
@@ -73,16 +91,29 @@ namespace MvcTomato.Controllers
         public ActionResult Edit(int id)
         {
             WorkingDay day = db.WorkingDays.Find(id);
+            if (day.OwnerId != User.Identity.GetUserId())
+            {
+                return View("Error");
+            }
             return View(day);
         }
 
         [HttpPost]
         public ActionResult Edit(WorkingDay day)
         {
-            db.Entry(day).State = EntityState.Modified;
-            if (day.Enter != null && day.Exit != null && day.DinnerStart != null && day.DinnerFinish != null)
+            var dayModel = db.WorkingDays.Find(day.Id);
+            if (dayModel.OwnerId != User.Identity.GetUserId())
             {
-                day.Finished = true;
+                return View("Error");
+            }
+            //db.Entry(day).State = EntityState.Modified;
+            dayModel.Enter = day.Enter;
+            dayModel.Exit = day.Exit;
+            dayModel.DinnerStart = day.DinnerStart;
+            dayModel.DinnerFinish = day.DinnerFinish;
+            if (dayModel.Enter != null && dayModel.Exit != null && dayModel.DinnerStart != null && dayModel.DinnerFinish != null)
+            {
+                dayModel.Finished = true;
             }
             db.SaveChanges();
             return RedirectToAction("Index");
@@ -90,8 +121,10 @@ namespace MvcTomato.Controllers
 
         public ActionResult History()
         {
+            string userId = User.Identity.GetUserId();
             // TODO: Decide how to show uncompeted days
             var days = db.WorkingDays
+                .Where(d => d.OwnerId == userId)
                 .Where(d => d.Date.Month == DateTime.Today.Month)
                 .OrderBy(d => d.Date)
                 .ToList()
@@ -108,9 +141,15 @@ namespace MvcTomato.Controllers
 
         public ActionResult Delete(int id)
         {
+            string userId = User.Identity.GetUserId();
             try
             {
                 WorkingDay day = db.WorkingDays.Find(id);
+                if (day.OwnerId != userId)
+                {
+                    Console.Error.WriteLine($"The user '{userId}' is not allowed to delete '{day.Id}' entity");
+                    return View("Error");
+                }
                 db.WorkingDays.Remove(day);
                 db.SaveChanges();
             }
